DEV-120 refactor: rename `lib.background` to `lib.bg`
