+++ lua/terra-core/lib/bg.lua
